--- bundle/config/scripts.go
@@ -1,13 +1,0 @@
-package config
-
-type Command string
-type ScriptHook string
-
-const (
-	ScriptPreInit    ScriptHook = "preinit"
-	ScriptPostInit   ScriptHook = "postinit"
-	ScriptPreBuild   ScriptHook = "prebuild"
-	ScriptPostBuild  ScriptHook = "postbuild"
-	ScriptPreDeploy  ScriptHook = "predeploy"
-	ScriptPostDeploy ScriptHook = "postdeploy"
-)
